requirements.txt
pandas
openai
python-dotenv
faiss-cpu
openai
